Add support XRP's total chunks XC field #1
Added logic to send XC field as first field if more than one chunk will be sent.
Added logic to check for an XC field on receive and store the total of incoming chunks and the amount of chunks already received for that profile.
When last incoming chunk is received, makes sure both values are equal for evaluation in other add-ons.