Use new printable separator for logged messages
We use a separator of two of the same character, so we can escape the character in the body text and easily un-escape it later. The character ~ was picked as the less likely character to be manually typed by the player while still be printed correctly.
Escape fields before sending and unescape them when received.
Incremented LibMSP version number and protocol version number