Replace msp:LoadCache() with msp.callback.dataload
The new callback provides the name and newly-created msp.char table for a character, permitting an addon with cached data to pre-fill saved information prior to anything else interacting with the newly-created table. This permits on-the-fly loading of cached data, rather than all-at-once, saving both execution time at startup and memory.