Clean up msp:Update(), trigger appropriate callbacks
msp:Update(), since using CRCs, only serves to rebuild the tooltip cache and update our own profile in msp.char. As such, it seems better to trigger the appropriate callbacks for when the profile is updated, to save LibMSP users some headaches when the player is the one whose profile changed.