Distinguish constants/variables, all variables in msp table
This commit makes a harder distinguishment between constants (UPPERCASE) and variables (lowerCamelCase) in style, and places all items with variable states into the msp table to make it easier to update them when a new LibMSP is loaded.

A few exceptions were made for the stylistic concerns to retain API-compatibility.