Fix missing parameter in Process call
Not sure on the ramifications of this bug, but this issue has been
around ever since the Process function gained its third argument.

The net result is the call being made here was effectively a no-op
because the action would always be nothing due to the field name
being used, and isSafe would have been nil so no branches could
possibly be taken.