Remove assumption that unit fields won't change
This is incorrect in a few edge cases - the first is something that
LibMSP tries to account for where a neutral-faction Pandaren is created
and then selects a faction later on in the levelling process. MSP
updates the `msp.my` data automatically in this case, however any
other players that queried for the "GF" field before that change
occurred would never see the updated faction once selected.

The other edge case is where a character is deleted and recreated, in
this case basically any and all of the unit fields could change with the
GU (GUID) being a guarantee.

Odds are the only reason this was never noticed was because these fields
probably aren't used by many people, however rather than keep a broken
assumption around it's better to fix this before we forget about it
and move on only to then get bitten ourselves one day in the far future.

Realistically since these fields are seldom used, the impact of allowing
them to be re-requested as frequently as other fields shouldn't be a
problem - and in the grand scheme of things the data these fields would
otherwise transmit should be minimal at best, especially given the
common case of them not actually changing.