Ensure AddFieldsToTooltip only adds fields once
If AddFieldsToTooltip is called with any field names already registered
in the `ttList` table, the AddTTField function would insert the field
again into the list.

This would then manifest in a later Update call by and cause the
duplicated fields to appear twice in the canned `ttContents`/`ttCache`
responses needlessly.

Checking if the table contains the field prior to adding ensures we
don't duplicate the field and keeps the size of TT field responses down.