Expose CRC32 function
This is intended for completeness to fix a hole where currently any
addon that wants to use the dataload callback needs to store the field
values _and_ the version data.

Exposing a way to use the built-in CRC32 functionality allows users
of the dataload callback to not need to worry about storing version
data as it can be inferred from the values instead.